refactor ConnectionRequest
